+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Studio plugin
 
+## Unreleased
+
+- Added <podcast:alternateEnclosure> support.
+
 ## 0.11.0 - 2023-07-03
 
 - Added <podcast:trailer> support ([#15](https://github.com/vnali/craft-studio/discussions/15)).